refactor(interactive): isolate figure composer workflows
